Create wikipedia_unreliable_sources.json
JSON file containing wikipedia's list of perennial sources with status "Generally unreliable", "Deprecated", and "Blacklisted". This is meant to replace to the simple lists of row IDs, which will not match domains due to underscores. Also, the Use column contains multiple domains associated with each source, which is more comprehensive.